+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 ---------
 
 #### Next Release
+- Supported query of multiple keywords.
 
 #### v1.15.1 (2016-11-07)
 - Updated deps: